# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Billing Service API",
    "version": "0.1.0",
    "description": "Manages billing accounts, credit balance, deductions, and Stripe integration for the Conductor platform."
  },
  "servers": [
    {
      "url": "http://localhost:3012"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "BalanceResponse": {
        "type": "object",
        "properties": {
          "balance_cents": {
            "type": "string",
            "description": "Spendable funds: credited_cents minus usage_cents. Includes provisioned holds, so this is the safety value for authorization, depletion, runway, and top-up checks."
          },
          "actual_balance_cents": {
            "type": "string",
            "description": "User-facing credit balance: credited funds minus actualized platform usage only. Provisioned holds are not subtracted here because they may later actualize or cancel."
          },
          "depleted": {
            "type": "boolean"
          }
        },
        "required": [
          "balance_cents",
          "actual_balance_cents",
          "depleted"
        ]
      },
...
      "InternalAccountTeardownDeletedRows": {
        "type": "object",
        "properties": {
          "billingAccounts": {
            "type": "integer"
          },
          "localPromos": {
            "type": "integer"
          },
          "creditDepletionEpisodes": {
            "type": "integer"
          },
          "campaignAuthorizeCosts": {
            "type": "integer"
          },
          "brandDailyBudgets": {
            "type": "integer"
          },
          "welcomeCreditClaims": {
            "type": "integer"
          }
        },
        "required": [
          "billingAccounts",
          "localPromos",
          "creditDepletionEpisodes",
          "campaignAuthorizeCosts",
          "brandDailyBudgets",
          "welcomeCreditClaims"
        ]
      },
      "InternalAccountTeardownResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "orgId": {
            "type": "string",
            "format": "uuid"
          },
          "deletedRows": {
            "$ref": "#/components/schemas/InternalAccountTeardownDeletedRows"
          }
        },
        "required": [
          "ok",
          "orgId",
          "deletedRows"
        ]
      },
...
      "DunningTickResponse": {
        "type": "object",
        "properties": {
          "processed": {
            "type": "integer"
          },
          "recovered": {
            "type": "integer"
          },
          "followup3dSent": {
            "type": "integer"
          },
          "followup10dSent": {
            "type": "integer"
          }
        },
        "required": [
          "processed",
          "recovered",
          "followup3dSent",
          "followup10dSent"
        ]
      },
      "CampaignAffordability": {
        "type": "object",
        "properties": {
          "affordable": {
            "type": "boolean"
          },
          "balanceCents": {
            "type": "string"
          },
          "lastRequiredCents": {
            "type": "string",
            "nullable": true
          },
          "hasHistory": {
            "type": "boolean"
          }
        },
        "required": [
          "affordable",
          "balanceCents",
          "lastRequiredCents",
          "hasHistory"
        ]
      },
...
      "TransferBrandTableResult": {
        "type": "object",
        "properties": {
          "tableName": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "tableName",
          "count"
        ]
      },
      "TransferBrandResponse": {
        "type": "object",
        "properties": {
          "updatedTables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferBrandTableResult"
            }
          }
        },
        "required": [
          "updatedTables"
        ]
      },
      "TransferBrandRequest": {
        "type": "object",
        "properties": {
          "sourceBrandId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceOrgId": {
            "type": "string",
            "format": "uuid"
          },
          "targetOrgId": {
            "type": "string",
            "format": "uuid"
          },
          "targetBrandId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "sourceBrandId",
          "sourceOrgId",
          "targetOrgId"
        ]
      }
    },
...
  },
  "paths": {
...
    "/internal/accounts/by-org/{orgId}": {
      "delete": {
        "summary": "Remove billing-owned state for a deleted org",
        "description": "Client-service cascade teardown leg for an internal org UUID. Removes only billing-service-owned org-scoped rows that can keep active billing effects alive: account topup config, local promo credits, dunning episodes, campaign affordability estimates, brand daily budgets, and welcome-credit claims. No cross-service fan-out. Idempotent: no rows for the org is still success.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Billing-owned org state removed; all counts may be zero on retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalAccountTeardownResponse"
                }
              }
            }
          },
          "400": {
            "description": "orgId is not a valid UUID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Database operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/dunning/tick": {
      "post": {
        "summary": "Run one out-of-credit dunning scheduler pass (ops / manual trigger)",
        "description": "Processes every open depletion episode: closes those whose balance was restored (stop-on-recharge, no email) and sends due +3d / +10d follow-ups. The same pass runs automatically on the in-process hourly scheduler; this route is for ops and testing. Idempotent — re-running never double-sends a stage.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Tick summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DunningTickResponse"
                }
              }
            }
          },
          "502": {
            "description": "Tick failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/campaigns/{campaignId}/affordability": {
      "get": {
        "summary": "Read-only pre-flight: can this org afford another run of campaign X?",
        "description": "Answers campaign-service's affordability question WITHOUT charging or reloading. Zero side effects — no charge, no reload, no depletion-episode mutation. Estimates the next run's cost as the required_cents of the campaign's LAST authorize attempt (a campaign re-runs the same workflow → ~constant cost). hasHistory=false (no authorize recorded yet) → affordable=true so a brand-new campaign can run once to establish its cost. Otherwise affordable = live balance >= lastRequiredCents.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "campaignId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Affordability verdict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignAffordability"
                }
              }
            }
          },
          "400": {
            "description": "campaignId is not a valid UUID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "stripe-service or runs-service unavailable (balance compose failed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/accounts/by-org/{orgId}/balance": {
      "get": {
        "summary": "User-less spendable balance for an org (platform/fleet reads)",
        "description": "Same spendable-balance snapshot as GET /v1/accounts/balance (balance_cents = credited − committed usage; actual_balance_cents = credited − actualized usage; depleted = balance_cents <= 0), but keyed by the orgId PATH param and callable with the service x-api-key ONLY — no x-org-id / x-user-id / x-run-id, no sentinel identity. For platform/staff fleet aggregators (accounts audit, send-forecast) that have no end-user in context. Pure read: no auto-reload, no depletion mutation. 404 when the org has no billing account; 502 when stripe-service/runs-service is unreachable.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Balance info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "orgId is not a valid UUID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Billing account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "stripe-service or runs-service unavailable (balance compose failed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/transfer-brand": {
      "post": {
        "summary": "Transfer all solo-brand rows from one org to another (billing + stripe-service)",
        "description": "Updates local_promos in billing AND proxies to stripe-service for ledger rows. Skips co-branding rows. Idempotent.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferBrandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer result with per-table update counts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferBrandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "stripe-service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}